+++ src-docs/src/views/guidelines/colors/core_palette.js
@@ -1,0 +1,56 @@
+import React from 'react';
+import { getSassVars } from '../_get_sass_vars';
+
+import {
+  EuiFlexGroup,
+  EuiFlexItem,
+  EuiIcon,
+  EuiCopy,
+} from '../../../../../src/components';
+import { rgbToHex } from '../../../../../src/services';
+import { scrollToSelector } from '../../../components/guide_page/guide_page_chrome';
+
+export const CorePalette = ({ theme, colors }) => {
+  const palette = getSassVars(theme);
+
+  function renderPaletteColor(palette, color, index) {
+    const hex = palette[color];
+    const name = color;
+
+    return (
+      <EuiFlexItem key={index} grow={false}>
+        <EuiCopy
+          title={`$${color}:
+          ${rgbToHex(hex.rgba).toUpperCase()}`}
+          beforeMessage={
+            <small>
+              <kbd>Click</kbd> to copy color name
+              <br />
+              <kbd>Shift + Click</kbd> to scroll to section
+            </small>
+          }
+          afterMessage={<small>Copied!</small>}
+          textToCopy={name}>
+          {copy => (
+            <EuiIcon
+              onClick={e => {
+                e.shiftKey ? scrollToSelector(`#${color}`) : copy();
+              }}
+              size="xxl"
+              type="stopFilled"
+              color={rgbToHex(hex.rgba)}
+            />
+          )}
+        </EuiCopy>
+      </EuiFlexItem>
+    );
+  }
+
+  return (
+    <EuiFlexGroup className="guideSection__shadedBox" gutterSize="s" wrap>
+      {colors.map(function(color, index) {
+        return renderPaletteColor(palette, color, index);
+      })}
+    </EuiFlexGroup>
+  );
+};
